renamed content.tsx to _content.tsx to stop docusaurus from treating it like a page instead of a compoment
